pyproject.toml

[project]
name = "bamboo"
version = "0.1.0"
description = "A CMS optimized for conference hosting"
authors = [
    {name = "Bamboo Team", email = "me@frostming.com"},
]
dependencies = [
    "apiflask>=2.1.0",
    "flask-sqlalchemy>=3.1.1",
    "flask-migrate>=4.0.5",
    "python-dotenv>=1.0.0",
    "python-jose[cryptography]>=3.3.0",
    "Flask-RQ2>=18.3",
    "Pillow>=10.1.0",
    "setuptools>=69.0.3",
    "psycopg2-binary>=2.9.9",
    "httpx>=0.27.0",
    "Frozen-Flask>=1.0.2",
]
requires-python = ">=3.12"
license = {text = "BSD 3-Clause"}

[tool.pdm]
distribution = false

[tool.pdm.scripts]
serve = { cmd = "flask run", help = "Run the development server" }
create-tables = { cmd = "flask create-tables", help = "Create tables" }
drop-tables = { cmd = "flask drop-tables", help = "Drop tables" }
worker = { cmd = "flask rq worker", help = "Run the RQ worker" }

[tool.pdm.dev-dependencies]
test = [
    "pytest>=7.4.3",
    "pytest-mock>=3.12.0",
    "fakeredis>=2.20.1",
    "pytest-httpx>=0.30.0",
]
dev = [
    "pre-commit>=3.6.0",
    "flask-shell-ipython>=0.5.1",
]

[tool.ruff]
line-length = 100
extend-select = [
  "I",    # isort
  "B",    # flake8-bugbear
  "C4",   # flake8-comprehensions
  "PGH",  # pygrep-hooks
  "RUF",  # ruff
  "W",    # pycodestyle
  "YTT",  # flake8-2020
]
extend-ignore = ["B018", "B019"]
src = ["src"]
exclude = ["tests/fixtures"]
target-version = "py312"

[tool.ruff.mccabe]
max-complexity = 10

[tool.ruff.isort]
known-first-party = ["bamboo"]

[tool.pytest.ini_options]
filterwarnings = [
  "ignore:pkg_resources is deprecated:DeprecationWarning",
  "ignore:datetime.datetime.utc:DeprecationWarning",
]
